+++ examples/test_libmeshb_p2_sol.c
@@ -21,7 +21,7 @@ int main()
       return(1);
 
    // Read the number vertices and associated solution size for memory allocation
-   NmbSol = GmfStatKwd(InpMsh, GmfHOSolAtEdgesP2, &NmbTyp, &SolSiz, TypTab, &deg, &NmbNod);
+   NmbSol = (int)GmfStatKwd(InpMsh, GmfHOSolAtEdgesP2, &NmbTyp, &SolSiz, TypTab, &deg, &NmbNod);
    printf("NmbSol = %d, NmbTyp = %d, SolSiz = %d, degree = %d, NmbNod = %d\n",
             NmbSol, NmbTyp, SolSiz, deg, NmbNod);
    SolTab = malloc( (NmbSol+1) * SolSiz * sizeof(double));
